+++ core/CHANGELOG.md
@@ -11,9 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add type `UnknownVariant` to represent an unknown enum variant. `ValueSerializer` and
   `ValueDeserializer` have been extended as well.
+- Add type `UnknownFields` to represent the unknown fields of a struct. `ValueSerializer` and
+  `ValueDeserializer` have been extended as well.
 - Enum introspection (`introspection::Enum`) can now specify a special variant as a fallback.
 - `ValueSerializer` and `ValueDeserializer` have been extended to support structs with a fallback
-  field of type `HashMap<u32, SerializedValue>`.
+  field of type `UnknownFields`.
 - Struct introspection (`introspection::Struct`) can now specify a special field as a fallback.
 
 ### Changed